admin-api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
bcrypt==4.3.0
PyJWT==2.10.1
